CmuxMobileCloudUI: present the Cloud flow full screen with its own navigation stack
Pushing the catalog inside the host Computers stack fired the section's
onDisappear before the catalog's onAppear, which dropped the tunnel lease to
zero, tore the tunnel down, and re-rendered the section, popping the push
(observed in the simulator on three builds). CloudFlowView now owns a
NavigationStack presented full screen from the Cloud row: the tunnel lifecycle
hangs off the stack container, which is stable across pushes, and every
destination registers at that stack's root. HIG: Modality, Navigation bars.
